% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{greeting_n_times}
\alias{greeting_n_times}
\title{Return string \code{"Hello world!"} n times to R.}
\usage{
greeting_n_times(n)
}
\description{
Return string \code{"Hello world!"} n times to R.
}
