% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{package}
\name{rpackageUsingRust-package}
\alias{rpackageUsingRust}
\alias{rpackageUsingRust-package}
\title{rpackageUsingRust: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/S3r4f1n/rpackageUsingRust}
  \item Report bugs at \url{https://github.com/S3r4f1n/rpackageUsingRust/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
